fixed reload page token

+++ backend/routes/categories.js
@@ -8,7 +8,8 @@ const router = express.Router();
 router.post("", checkAuth, (req, res, next) => {
     const category = new Category({
         name: req.body.name,
-        type: [req.body.type]
+        type: [req.body.type],
+        creator: req.userData.userId
     });
     category.save().then( crearedCategory => {
         res.status(201).json({
@@ -21,7 +22,8 @@ router.post("", checkAuth, (req, res, next) => {
 router.put("/:id", checkAuth, (req, res, next) => {
     const category = new Category({
         name: req.body.name,
-        type: req.body.type
+        type: req.body.type,
+        creator: req.userData.userId
     });
     category.updateOne({ _id: req.params.id }, category).then( result => {
         res.status(200).json({
@@ -49,9 +51,8 @@ router.get("/:id",(req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Category.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Category deleted!"})
+    Category.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        res.status(200).json({ message: "Category deleted!" })
     });
 });
